+++ app/models/gaggle/goose.rb
@@ -13,7 +13,7 @@ module Gaggle
 
     validates :name, presence: true, uniqueness: true
 
-    def private_thread(goose)
+    def private_channel(goose)
       Gaggle::Message.where(messageable: self, goose:).or(Gaggle::Message.where(messageable: goose, goose: self))
     end
 
@@ -30,33 +30,9 @@ module Gaggle
 
         #{`rake -D gaggle:`}"
 
-        1. To get a list of thread names and their IDs:
-           bin/rails gaggle:get_threads
-
-        2. To retrieve a specific thread and its messages:
-           bin/rails gaggle:get_thread_messages thread_id={id}
-           Replace {id} with the actual thread ID.
-
-        3. To post a message in a thread:
-           bin/rails gaggle:send_message thread_id={id} content="{content}"
-           Replace {id} with thread id and {content} with the content of the message
-           BE SURE to escape anything that may break out of the surrounding quotation marks. It needs to all be delivered as a single argument.
-
-        4. To create a new thread:
-           bin/rails gaggle:create_thread name="{name}"
-           Replace {name} with the name of the thread
-
-        4. To update the name of a thread:
-           bin/rails gaggle:update_thread thread_id={id} name="{name}"
-           Replace {id} with thread id and {name} with the new name
-
-        5. To delete a thread:
-           bin/rails gaggle:delete_thread thread_id={id}
-           Replace {id} with thread id
-
         When you need to communicate:
-        1. First, check existing threads
-        2. If a relevant thread exists, use its ID to view messages, otherwise create a new one
+        1. First, check existing channels
+        2. If a relevant channel exists, use its ID to view messages, otherwise create a new one
         3. If you receive a notification, consider replying to the message
 
         BE ABSOLUTELY SURE to receive a successful response after sending a new message. If you do not, your formatting may be wrong.
@@ -66,7 +42,7 @@ module Gaggle
 
     def begin_prompt
       <<~TEXT
-        Begin by reading the available threads and responding where necessary
+        Begin by reading the available channels and responding where necessary
       TEXT
     end