+++ app/models/gaggle/session.rb
@@ -11,7 +11,7 @@ module Gaggle
       @@running_executables[to_global_id] = ::Thread.new do
         Rails.logger.info "Starting executable for: #{goose.name}"
         Open3.popen3({ "GOOSE_ID" => goose.id.to_s }, "goose session") do |stdin, stdout, stderr, wait_thr|
-          ::Thread.current[:stdin] = stdin
+          Thread.current[:stdin] = stdin
           @stdout = stdout
           @stderr = stderr
           @pid = wait_thr.pid
@@ -20,29 +20,29 @@ module Gaggle
           log_file_path = Rails.root.join(self.log_file)
           FileUtils.mkdir_p(File.dirname(log_file_path))
           FileUtils.touch(log_file_path)
-          ::Thread.current[:logger] = Logger.new(log_file_path)
-          ::Thread.current[:logger].info "Session started"
+          Thread.current[:logger] = Logger.new(log_file_path)
+          Thread.current[:logger].info "Session started"
           save!
 
           begin
             while (line = @stdout.gets)
-              ::Thread.current[:logger].info line
+              Thread.current[:logger].info line
               broadcast_append target: dom_id(self, :code), content: Strings::ANSI.sanitize(line)
             end
           rescue IOError => e
-            ::Thread.current[:logger].error "Output stream closed: #{e.message}"
+            Thread.current[:logger].error "Output stream closed: #{e.message}"
           end
         end
       end
     end
 
     def write_to_executable(input)
-      if thread = @@running_executables[to_global_id]
+      if channel = @@running_executables[to_global_id]
 
 
-        thread[:logger].info "Sending input: #{input}"
+        channel[:logger].info "Sending input: #{input}"
         broadcast_append target: dom_id(self, :code), content: Strings::ANSI.sanitize(input)
-        thread[:stdin].write(input.gsub("\n", "__NEWLINE_PLACEHOLDER__") + "\n")
+        channel[:stdin].write(input.gsub("\n", "__NEWLINE_PLACEHOLDER__") + "\n")
       else
         Rails.logger.error "No running executable found for session #{to_global_id}"
       end