+++ db/migrate_gaggle/20250218170424_rename_output_to_log_file_gaggle_sessions.rb
@@ -1,6 +1,0 @@
-class RenameOutputToLogFileGaggleSessions < ActiveRecord::Migration[8.0]
-  def change
-    rename_column :gaggle_sessions, :output, :log_file
-    change_column :gaggle_sessions, :log_file, :string
-  end
-end