+++ db/migrate_gaggle/20250218235005_add_thread_id_to_gaggle_notifications.rb
@@ -1,5 +1,0 @@
-class AddThreadIdToGaggleNotifications < ActiveRecord::Migration[6.0]
-  def change
-    add_reference :gaggle_notifications, :messageable, polymorphic: true, null: true
-  end
-end