+++ db/migrate_gaggle/20250219032639_replace_thread_with_messageable_message.rb
@@ -1,5 +1,0 @@
-class ReplaceThreadWithMessageableMessage < ActiveRecord::Migration[8.0]
-  def change
-    add_reference :gaggle_messages, :messageable, polymorphic: true, null: true
-  end
-end