+++ test/dummy/config/puma.rb
@@ -3,29 +3,29 @@
 # about methods provided by the DSL, see https://puma.io/puma/Puma/DSL.html.
 #
 # Puma starts a configurable number of processes (workers) and each process
-# serves each request in a thread from an internal thread pool.
+# serves each request in a channel from an internal channel pool.
 #
 # You can control the number of workers using ENV["WEB_CONCURRENCY"]. You
 # should only set this value when you want to run 2 or more workers. The
 # default is already 1.
 #
-# The ideal number of threads per worker depends both on how much time the
+# The ideal number of channels per worker depends both on how much time the
 # application spends waiting for IO operations and on how much you wish to
 # prioritize throughput over latency.
 #
-# As a rule of thumb, increasing the number of threads will increase how much
+# As a rule of thumb, increasing the number of channels will increase how much
 # traffic a given process can handle (throughput), but due to CRuby's
 # Global VM Lock (GVL) it has diminishing returns and will degrade the
 # response time (latency) of the application.
 #
-# The default is set to 3 threads as it's deemed a decent compromise between
+# The default is set to 3 channels as it's deemed a decent compromise between
 # throughput and latency for the average Rails application.
 #
 # Any libraries that use a connection pool or another resource pool should
 # be configured to provide at least as many connections as the number of
-# threads. This includes Active Record's `pool` parameter in `database.yml`.
-threads_count = ENV.fetch("RAILS_MAX_THREADS", 3)
-threads threads_count, threads_count
+# channels. This includes Active Record's `pool` parameter in `database.yml`.
+channels_count = ENV.fetch("RAILS_MAX_THREADS", 3)
+channels channels_count, channels_count
 
 # Specifies the `port` that Puma will listen on to receive requests; default is 3000.
 port ENV.fetch("PORT", 3000)