+++ test/dummy/db/schema.rb
@@ -21,12 +21,12 @@ ActiveRecord::Schema[8.0].define(version: 2025_02_14_180303) do
 
   create_table "gaggle_messages", force: :cascade do |t|
     t.text "content", null: false
-    t.integer "gaggle_thread_id", null: false
+    t.integer "gaggle_channel_id", null: false
     t.integer "gaggle_goose_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.index ["gaggle_goose_id"], name: "index_gaggle_messages_on_gaggle_goose_id"
-    t.index ["gaggle_thread_id"], name: "index_gaggle_messages_on_gaggle_thread_id"
+    t.index [ "gaggle_goose_id" ], name: "index_gaggle_messages_on_gaggle_goose_id"
+    t.index [ "gaggle_channel_id" ], name: "index_gaggle_messages_on_gaggle_channel_id"
   end
 
   create_table "gaggle_notifications", force: :cascade do |t|
@@ -35,18 +35,18 @@ ActiveRecord::Schema[8.0].define(version: 2025_02_14_180303) do
     t.datetime "read_at", precision: nil
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.index ["gaggle_goose_id"], name: "index_gaggle_notifications_on_gaggle_goose_id"
-    t.index ["gaggle_message_id"], name: "index_gaggle_notifications_on_gaggle_message_id"
+    t.index [ "gaggle_goose_id" ], name: "index_gaggle_notifications_on_gaggle_goose_id"
+    t.index [ "gaggle_message_id" ], name: "index_gaggle_notifications_on_gaggle_message_id"
   end
 
-  create_table "gaggle_threads", force: :cascade do |t|
+  create_table "gaggle_channels", force: :cascade do |t|
     t.string "name", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
 
   add_foreign_key "gaggle_messages", "gaggle_gooses"
-  add_foreign_key "gaggle_messages", "gaggle_threads"
+  add_foreign_key "gaggle_messages", "gaggle_channels"
   add_foreign_key "gaggle_notifications", "gaggle_gooses"
   add_foreign_key "gaggle_notifications", "gaggle_messages"
 end